perf(usage): speed up model scan and split helpers
